make parameter printing more consistent via lambda function:
  - New lambda function that is used to print (almost) all MDFlexOptions
  - Add std::set to types for which autopas::utils::ArrayUtils::is_container() returns true
    so that autopas::utils::ArrayUtils::operator<< works for std::set
  - Add type traits to detect if a type is a smart-pointer
  - change type of MDFlexConfig::valueOffset to int to avoid casting.

+++ examples/md-flexible/src/configuration/MDFlexConfig.h
@@ -645,7 +645,7 @@ class MDFlexConfig {
   /**
    * valueOffset used for cli-output alignment
    */
-  static constexpr size_t valueOffset{33};
+  static constexpr int valueOffset{33};
 
  private:
   /**
